+++ src/analysis/Interpreter.jl
@@ -3,7 +3,7 @@ module Interpreter
 export LSInterpreter
 
 using JuliaSyntax: JuliaSyntax as JS
-using JET: CC, JET
+using JET: CC, JET, JuliaInterpreter
 using ..JETLS:
     AnalysisRequest, AnalysisResult, SavedFileInfo, Server, JETLS, JETLS_DEV_MODE,
     is_cancelled, send_progress, yield_to_endpoint
@@ -24,18 +24,25 @@ struct LSInterpreter{S<:Server} <: JET.ConcreteInterpreter
     analyzer::LSAnalyzer
     counter::Counter
     activation_done::Union{Nothing,Base.Event}
+    warning_reports::Vector{JETLS.ToplevelWarningReport}
+    current_node::Base.RefValue{JS.SyntaxNode}
     state::JET.InterpretationState
     function LSInterpreter(
             server::S, request::AnalysisRequest, analyzer::LSAnalyzer, counter::Counter,
             activation_done::Union{Nothing,Base.Event}
         ) where S<:Server
-        return new{S}(server, request, analyzer, counter, activation_done)
+        return new{S}(server, request, analyzer, counter, activation_done,
+            JETLS.ToplevelWarningReport[], Base.RefValue{JS.SyntaxNode}())
     end
     function LSInterpreter(
             server::S, request::AnalysisRequest, analyzer::LSAnalyzer, counter::Counter,
-            activation_done::Union{Nothing,Base.Event}, state::JET.InterpretationState
+            activation_done::Union{Nothing,Base.Event},
+            warning_reports::Vector{JETLS.ToplevelWarningReport},
+            current_node::Base.RefValue{JS.SyntaxNode},
+            state::JET.InterpretationState,
         ) where S<:Server
-        return new{S}(server, request, analyzer, counter, activation_done, state)
+        return new{S}(server, request, analyzer, counter, activation_done,
+            warning_reports, current_node, state)
     end
 end
 
@@ -52,7 +59,7 @@ JET.InterpretationState(interp::LSInterpreter) = interp.state
 function JET.ConcreteInterpreter(interp::LSInterpreter, state::JET.InterpretationState)
     return LSInterpreter(
         interp.server, interp.request, interp.analyzer, interp.counter,
-        interp.activation_done, state)
+        interp.activation_done, interp.warning_reports, interp.current_node, state)
 end
 JET.ToplevelAbstractAnalyzer(interp::LSInterpreter) = interp.analyzer
 function JET.ToplevelAbstractAnalyzer(
@@ -70,6 +77,27 @@ end
 
 # overloads
 # =========
+
+struct MethodDefinitionInfo
+    filename::String
+    mod::Module
+    src::Core.CodeInfo
+end
+
+const empty_lines_range = typemax(Int) => typemin(Int)
+
+function get_lines_in_src(filepath::AbstractString, src::Core.CodeInfo)
+    lines = empty_lines_range
+    for pc in 1:length(src.code)
+        lins = Base.IRShow.buildLineInfoNode(src.debuginfo, nothing, pc)
+        for lin in lins
+            if filepath == String(lin.file)
+                lines = min(lin.line,first(lines)) => max(lin.line,last(lines))
+            end
+        end
+    end
+    return lines
+end
 
 mutable struct SignatureAnalysisProgress
     const reports::Vector{JET.InferenceErrorReport}
@@ -89,7 +117,7 @@ end
 
 function cache_intermediate_analysis_result!(interp::LSInterpreter)
     result = JET.JETToplevelResult(interp.analyzer, interp.state.res, "LSInterpreter (intermediate result)", ())
-    intermediate_result = JETLS.new_analysis_result(interp.request, result)
+    intermediate_result = JETLS.new_analysis_result(interp, interp.request, result)
     JETLS.update_analysis_cache!(interp.server.state.analysis_manager, intermediate_result)
 end
 
@@ -108,8 +136,23 @@ function JET.analyze_from_definitions!(interp::LSInterpreter, config::JET.Toplev
     res = JET.InterpretationState(interp).res
     reset_report_target_modules!(interp.analyzer, res.analyzed_files)
 
+    # Detect method overwrites
+    seen_sigs = IdDict{Type,MethodDefinitionInfo}()
+    for signature_info in res.signature_infos
+        (; filename, mod, tt, src) = signature_info
+        if haskey(seen_sigs, tt)
+            lines = get_lines_in_src(filename, src)
+            original_definition = seen_sigs[tt]
+            original_filename = original_definition.filename
+            original_lines = get_lines_in_src(original_filename, original_definition.src)
+            push!(interp.warning_reports, JETLS.MethodOverwriteReport(mod, tt, filename, lines, original_filename, original_lines))
+        else
+            seen_sigs[tt] = MethodDefinitionInfo(filename, mod, src)
+        end
+    end
+
     entrypoint = config.analyze_from_definitions
-    n_sigs = length(res.toplevel_signatures)
+    n_sigs = length(res.signature_infos)
     n_sigs == 0 && return
     cancellable_token = interp.request.cancellable_token
     if cancellable_token !== nothing
@@ -131,7 +174,7 @@ function JET.analyze_from_definitions!(interp::LSInterpreter, config::JET.Toplev
             if cancellable_token !== nothing && is_cancelled(cancellable_token.cancel_flag)
                 return
             end
-            tt = res.toplevel_signatures[i]
+            (; tt) = res.signature_infos[i]
             # Create a new analyzer with fresh local caches (`inf_cache` and `analysis_results`)
             # to avoid data races between concurrent signature analysis tasks
             analyzer = JET.ToplevelAbstractAnalyzer(interp, JET.non_toplevel_concretized;
@@ -212,7 +255,7 @@ function JET.virtual_process!(interp::LSInterpreter,
     return res
 end
 
-function JET.try_read_file(interp::LSInterpreter, include_context::Module, filename::AbstractString)
+function JET.try_read_file(interp::LSInterpreter, _include_context::Module, filename::AbstractString)
     uri = filename2uri(filename)
     fi = JETLS.get_saved_file_info(interp.server.state, uri)
     if !isnothing(fi)
@@ -225,6 +268,72 @@ function JET.try_read_file(interp::LSInterpreter, include_context::Module, filen
     end
     # fallback to the default file-system-based include
     return read(filename, String)
+end
+
+# XXX This is an ad-hoc overload to make `node` available in the following `JuliaInterpreter.step_expr!`
+function JET.lower_with_err_handling(interp::LSInterpreter, node::JS.SyntaxNode, xblk::Expr)
+    interp.current_node[] = node
+    @invoke JET.lower_with_err_handling(interp::JET.ConcreteInterpreter, node::JS.SyntaxNode, xblk::Expr)
+end
+
+function JuliaInterpreter.step_expr!(
+        interp::LSInterpreter, frame::JuliaInterpreter.Frame, @nospecialize(node),
+        istoplevel::Bool
+    )
+    if Meta.isexpr(node, :call) && length(node.args) ≥ 4
+        func = JuliaInterpreter.lookup(frame, node.args[1])
+        if func === Core._typebody!
+            structtyp = JuliaInterpreter.lookup(frame, node.args[3])
+            if structtyp isa Type
+                ftypes = JuliaInterpreter.lookup(frame, node.args[4])::Core.SimpleVector
+                fnames = fieldnames(structtyp)
+                for (fname, ft) in zip(fnames, ftypes)
+                    if JETLS.is_abstract_fieldtype(ft)
+                        filename = JET.InterpretationState(interp).filename
+                        fieldline = try_extract_field_line(interp, frame, nameof(structtyp), fname)
+                        push!(interp.warning_reports, JETLS.AbstractFieldReport(filename, fieldline, structtyp, fname, ft))
+                    end
+                end
+            end
+        end
+    end
+    @invoke JuliaInterpreter.step_expr!(
+        interp::JET.ConcreteInterpreter, frame::JuliaInterpreter.Frame, node::Any,
+        istoplevel::Bool
+    )
+end
+
+# TODO Use lowered `SyntaxTree` for finding field line for macro-generated structs
+function try_extract_field_line(interp::LSInterpreter, frame::JuliaInterpreter.Frame, structname::Symbol, fname::Symbol)
+    isassigned(interp.current_node) || return JuliaInterpreter.linenumber(frame)
+    node = interp.current_node[]
+    return @something _try_extract_field_line(node, structname, fname) return JuliaInterpreter.linenumber(frame)
+end
+
+function _try_extract_field_line(node::JS.SyntaxNode, structname::Symbol, fname::Symbol)
+    if JS.kind(node) === JS.K"struct" && JS.numchildren(node) ≥ 2
+        structnm = node[1]
+        if JS.kind(structnm) === JS.K"curly" && JS.numchildren(structnm) ≥ 1
+            structnm = structnm[1]
+        end
+        if (let data = structnm.data; data !== nothing && data.val === structname; end)
+            for i = 1:JS.numchildren(node[2])
+                field = node[2][i]
+                if (JS.kind(field) === JS.K"::" && JS.numchildren(field) ≥ 1 &&
+                    let data = field[1].data; data !== nothing && data.val === fname; end)
+                    return field
+                elseif let data = field.data; data !== nothing && data.val === fname; end
+                    return field
+                end
+            end
+        end
+        return nothing
+    else
+        for i = 1:JS.numchildren(node)
+            return @something _try_extract_field_line(node[i], structname, fname) continue
+        end
+        return nothing
+    end
 end
 
 end # module Interpreter